+++ R/plsrannar.R
@@ -1,4 +1,4 @@
-plsrannar <- function(X, Y, weights = NULL, nlv) {
+plsrannar <- function(X, Y, scaling = "Centered", weights = NULL, nlv) {
     X <- .mat(X)
     Y <- .mat(Y, "y")     
     n <- dim(X)[1]
@@ -7,9 +7,29 @@ plsrannar <- function(X, Y, weights = NULL, nlv) {
         weights <- rep(1, n)
     weights <- .mweights(weights)
     xmeans <- .colmeans(X, weights = weights) 
-    X <- .center(X, xmeans)
-    ymeans <- .colmeans(Y, weights = weights) 
-    Y <- .center(Y, ymeans)
+    ymeans <- .colmeans(Y, weights = weights)
+    xsds <- sqrt(.colvars(X, weights = weights)*nrow(X)/(nrow(X)-1))
+    ysds <- sqrt(.colvars(Y, weights = weights)*nrow(Y)/(nrow(Y)-1))
+    # xsds <- sqrt(apply(X, MARGIN = 2, FUN= var))
+    # ysds <- sqrt(apply(Y, MARGIN = 2, FUN= var))
+    
+    if(scaling == "Centered"){
+      X <- .center(X, xmeans)
+      Y <- .center(Y, ymeans)
+    }
+    if(scaling == "Pareto"){
+      X <- .center(X, xmeans)
+      X <- scale(X, center = FALSE, scale = sqrt(xsds))
+      Y <- .center(Y, ymeans)
+      Y <- scale(Y, center = FALSE, scale = sqrt(ysds))
+    }
+    if(scaling == "CtReduced"){
+      X <- .center(X, xmeans)
+      X <- scale(X, center = FALSE, scale = xsds)
+      Y <- .center(Y, ymeans)
+      Y <- scale(Y, center = FALSE, scale = ysds)
+    }
+    
     nam <- paste("lv", seq_len(nlv), sep = "")
     U <- T <- Tclass <- matrix(nrow = n, ncol = nlv, 
                                dimnames = list(row.names(X), nam))                     
@@ -45,6 +65,6 @@ plsrannar <- function(X, Y, weights = NULL, nlv) {
     R <- W %*% solve(crossprod(P, W))
     structure(
         list(T = Tclass, P = P, R = R, W = W, C = C, TT = TT,
-            xmeans = xmeans, ymeans = ymeans, weights = weights, U = U),
+            xmeans = xmeans, ymeans = ymeans, xsds = xsds, ysds = ysds, weights = weights, scaling = scaling, U = U),
         class = c("Plsr", "Pls"))    
     }