+++ inst/NEWS.md
@@ -7,3 +7,4 @@
 ## **Version 0.1-2**
 
 - New functions :  vip, orthog, sopls, soplsrda, soplslds, soplsqda
+- Function changes : "scaling" argument added to plskern, plsrannar, plsnipals, plsrda, plslda, plsqda functions